Add error handling for not found cases in academicDepartment.service.ts and academicFaculty.service.ts

+++ src/app/modules/academicFaculty/academicFaculty.service.ts
@@ -1,3 +1,5 @@
+import httpStatus from 'http-status';
+import AppError from '../../errors/AppError';
 import { TAcademicFaculty } from './academicFaculty.interface';
 import { AcademicFaculty } from './academicFaculty.model';
 
@@ -12,7 +14,11 @@ const getAllAcademicFacultiesFromDB = async () => {
 };
 
 const getSingleAcademicFacultyFromDB = async (id: string) => {
+  // check if academic faculty exists
   const result = await AcademicFaculty.findById(id);
+  if (!result) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Academic Faculty not found');
+  }
   return result;
 };
 
@@ -20,6 +26,11 @@ const updateAcademicFacultyIntoDB = async (
   id: string,
   payload: Partial<TAcademicFaculty>,
 ) => {
+  // check if academic faculty exists
+  const academicFaculty = await AcademicFaculty.findById(id);
+  if (!academicFaculty) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Academic Faculty not found');
+  }
   const result = await AcademicFaculty.findByIdAndUpdate(id, payload, {
     new: true,
   });
@@ -27,6 +38,12 @@ const updateAcademicFacultyIntoDB = async (
 };
 
 const deleteAcademicFacultyFromDB = async (id: string) => {
+  // check if academic faculty exists
+  const academicFaculty = await AcademicFaculty.findById(id);
+  if (!academicFaculty) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Academic Faculty not found');
+  }
+
   const result = await AcademicFaculty.findByIdAndDelete(id);
   return result;
 };
